+++ src/__generated__/ArtworkContextArtistTestsQuery.graphql.ts
@@ -1,0 +1,590 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+import { ArtworkContextArtist_artwork$ref } from "./ArtworkContextArtist_artwork.graphql";
+export type ArtworkContextArtistTestsQueryVariables = {
+    readonly excludeArtworkIds?: ReadonlyArray<string | null> | null;
+};
+export type ArtworkContextArtistTestsQueryResponse = {
+    readonly artwork: ({
+        readonly " $fragmentRefs": ArtworkContextArtist_artwork$ref;
+    }) | null;
+};
+export type ArtworkContextArtistTestsQuery = {
+    readonly response: ArtworkContextArtistTestsQueryResponse;
+    readonly variables: ArtworkContextArtistTestsQueryVariables;
+};
+
+
+
+/*
+query ArtworkContextArtistTestsQuery(
+  $excludeArtworkIds: [String]
+) {
+  artwork(id: "abbas-kiarostami-untitled-7") {
+    ...ArtworkContextArtist_artwork
+    __id
+  }
+}
+
+fragment ArtworkContextArtist_artwork on Artwork {
+  ...ArtistArtworkGrid_artwork
+  __id
+}
+
+fragment ArtistArtworkGrid_artwork on Artwork {
+  id
+  artist {
+    name
+    artworks_connection(first: 8, sort: PUBLISHED_AT_DESC, exclude: $excludeArtworkIds) {
+      pageInfo {
+        startCursor
+        endCursor
+        hasPreviousPage
+        hasNextPage
+      }
+      edges {
+        node {
+          ...GenericGrid_artworks
+          __id
+          __typename
+        }
+        cursor
+      }
+    }
+    __id
+  }
+  __id
+}
+
+fragment GenericGrid_artworks on Artwork {
+  __id
+  id
+  image {
+    aspect_ratio
+  }
+  ...ArtworkGridItem_artwork
+}
+
+fragment ArtworkGridItem_artwork on Artwork {
+  title
+  date
+  sale_message
+  is_in_auction
+  is_biddable
+  is_acquireable
+  is_offerable
+  id
+  sale {
+    is_auction
+    is_live_open
+    is_open
+    is_closed
+    display_timely_at
+    __id
+  }
+  sale_artwork {
+    opening_bid {
+      display
+    }
+    current_bid {
+      display
+    }
+    bidder_positions_count
+    sale {
+      is_closed
+      __id
+    }
+    __id
+  }
+  image {
+    url(version: "large")
+    aspect_ratio
+  }
+  artists(shallow: true) {
+    name
+    __id
+  }
+  partner {
+    name
+    __id
+  }
+  href
+  __id
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "LocalArgument",
+    "name": "excludeArtworkIds",
+    "type": "[String]",
+    "defaultValue": null
+  }
+],
+v1 = [
+  {
+    "kind": "Literal",
+    "name": "id",
+    "value": "abbas-kiarostami-untitled-7",
+    "type": "String!"
+  }
+],
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+},
+v3 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "id",
+  "args": null,
+  "storageKey": null
+},
+v4 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "name",
+  "args": null,
+  "storageKey": null
+},
+v5 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "is_closed",
+  "args": null,
+  "storageKey": null
+},
+v6 = [
+  {
+    "kind": "ScalarField",
+    "alias": null,
+    "name": "display",
+    "args": null,
+    "storageKey": null
+  }
+],
+v7 = [
+  v4,
+  v2
+];
+return {
+  "kind": "Request",
+  "operationKind": "query",
+  "name": "ArtworkContextArtistTestsQuery",
+  "id": "794c5897ceac91e17f89e91abaeb0e9d",
+  "text": null,
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "ArtworkContextArtistTestsQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artwork",
+        "storageKey": "artwork(id:\"abbas-kiarostami-untitled-7\")",
+        "args": v1,
+        "concreteType": "Artwork",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "ArtworkContextArtist_artwork",
+            "args": null
+          },
+          v2
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "ArtworkContextArtistTestsQuery",
+    "argumentDefinitions": v0,
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "artwork",
+        "storageKey": "artwork(id:\"abbas-kiarostami-untitled-7\")",
+        "args": v1,
+        "concreteType": "Artwork",
+        "plural": false,
+        "selections": [
+          v3,
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "artist",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "Artist",
+            "plural": false,
+            "selections": [
+              v4,
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "artworks_connection",
+                "storageKey": null,
+                "args": [
+                  {
+                    "kind": "Variable",
+                    "name": "exclude",
+                    "variableName": "excludeArtworkIds",
+                    "type": "[String]"
+                  },
+                  {
+                    "kind": "Literal",
+                    "name": "first",
+                    "value": 8,
+                    "type": "Int"
+                  },
+                  {
+                    "kind": "Literal",
+                    "name": "sort",
+                    "value": "PUBLISHED_AT_DESC",
+                    "type": "ArtworkSorts"
+                  }
+                ],
+                "concreteType": "ArtworkConnection",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "LinkedField",
+                    "alias": null,
+                    "name": "pageInfo",
+                    "storageKey": null,
+                    "args": null,
+                    "concreteType": "PageInfo",
+                    "plural": false,
+                    "selections": [
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "startCursor",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "endCursor",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "hasPreviousPage",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "hasNextPage",
+                        "args": null,
+                        "storageKey": null
+                      }
+                    ]
+                  },
+                  {
+                    "kind": "LinkedField",
+                    "alias": null,
+                    "name": "edges",
+                    "storageKey": null,
+                    "args": null,
+                    "concreteType": "ArtworkEdge",
+                    "plural": true,
+                    "selections": [
+                      {
+                        "kind": "LinkedField",
+                        "alias": null,
+                        "name": "node",
+                        "storageKey": null,
+                        "args": null,
+                        "concreteType": "Artwork",
+                        "plural": false,
+                        "selections": [
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_acquireable",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          v2,
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "image",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "Image",
+                            "plural": false,
+                            "selections": [
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "aspect_ratio",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "url",
+                                "args": [
+                                  {
+                                    "kind": "Literal",
+                                    "name": "version",
+                                    "value": "large",
+                                    "type": "[String]"
+                                  }
+                                ],
+                                "storageKey": "url(version:\"large\")"
+                              }
+                            ]
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "title",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "date",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "sale_message",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_in_auction",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_biddable",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          v3,
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "is_offerable",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "sale",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "Sale",
+                            "plural": false,
+                            "selections": [
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "is_auction",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "is_live_open",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "is_open",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              v5,
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "display_timely_at",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              v2
+                            ]
+                          },
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "sale_artwork",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "SaleArtwork",
+                            "plural": false,
+                            "selections": [
+                              {
+                                "kind": "LinkedField",
+                                "alias": null,
+                                "name": "opening_bid",
+                                "storageKey": null,
+                                "args": null,
+                                "concreteType": "SaleArtworkOpeningBid",
+                                "plural": false,
+                                "selections": v6
+                              },
+                              {
+                                "kind": "LinkedField",
+                                "alias": null,
+                                "name": "current_bid",
+                                "storageKey": null,
+                                "args": null,
+                                "concreteType": "SaleArtworkCurrentBid",
+                                "plural": false,
+                                "selections": v6
+                              },
+                              {
+                                "kind": "ScalarField",
+                                "alias": null,
+                                "name": "bidder_positions_count",
+                                "args": null,
+                                "storageKey": null
+                              },
+                              {
+                                "kind": "LinkedField",
+                                "alias": null,
+                                "name": "sale",
+                                "storageKey": null,
+                                "args": null,
+                                "concreteType": "Sale",
+                                "plural": false,
+                                "selections": [
+                                  v5,
+                                  v2
+                                ]
+                              },
+                              v2
+                            ]
+                          },
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "artists",
+                            "storageKey": "artists(shallow:true)",
+                            "args": [
+                              {
+                                "kind": "Literal",
+                                "name": "shallow",
+                                "value": true,
+                                "type": "Boolean"
+                              }
+                            ],
+                            "concreteType": "Artist",
+                            "plural": true,
+                            "selections": v7
+                          },
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "partner",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "Partner",
+                            "plural": false,
+                            "selections": v7
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "href",
+                            "args": null,
+                            "storageKey": null
+                          },
+                          {
+                            "kind": "ScalarField",
+                            "alias": null,
+                            "name": "__typename",
+                            "args": null,
+                            "storageKey": null
+                          }
+                        ]
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "cursor",
+                        "args": null,
+                        "storageKey": null
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                "kind": "LinkedHandle",
+                "alias": null,
+                "name": "artworks_connection",
+                "args": [
+                  {
+                    "kind": "Variable",
+                    "name": "exclude",
+                    "variableName": "excludeArtworkIds",
+                    "type": "[String]"
+                  },
+                  {
+                    "kind": "Literal",
+                    "name": "first",
+                    "value": 8,
+                    "type": "Int"
+                  },
+                  {
+                    "kind": "Literal",
+                    "name": "sort",
+                    "value": "PUBLISHED_AT_DESC",
+                    "type": "ArtworkSorts"
+                  }
+                ],
+                "handle": "connection",
+                "key": "GenericGrid_artworks_connection",
+                "filters": [
+                  "sort",
+                  "exclude"
+                ]
+              },
+              v2
+            ]
+          },
+          v2
+        ]
+      }
+    ]
+  }
+};
+})();
+(node as any).hash = '260f998e90def0188b619ca86b5ff2e5';
+export default node;